+++ tests/Js2IL.Tests/Classes/ExecutionTests.cs
@@ -62,6 +62,7 @@ namespace Js2IL.Tests.Classes
         [Fact] public Task Classes_ClassMethod_CallsAnotherMethod() { var testName = nameof(Classes_ClassMethod_CallsAnotherMethod); return ExecutionTest(testName); }
         [Fact] public Task Classes_AccessorMethods_InstanceAndStatic() { var testName = nameof(Classes_AccessorMethods_InstanceAndStatic); return ExecutionTest(testName); }
         [Fact] public Task Classes_ClassMethod_ForLoop_CallsAnotherMethod() { var testName = nameof(Classes_ClassMethod_ForLoop_CallsAnotherMethod); return ExecutionTest(testName); }
+        [Fact] public Task Classes_ClassMethod_ParameterTypeInference_DirectThis_Number() { var testName = nameof(Classes_ClassMethod_ParameterTypeInference_DirectThis_Number); return ExecutionTest(testName); }
 
         // Test parameter destructuring in class methods
         [Fact]