+++ tests/Js2IL.Tests/Function/JavaScript/Function_ParameterTypeInference_EscapedFunction_KeepsObjectSignature.js
@@ -1,0 +1,9 @@
+"use strict";
+
+function add(a, b) {
+    return a + b;
+}
+
+module.exports = add;
+
+console.log(add(1, 2));